requirements.txt
python-dotenv
langchain
openai
langchain_community
